+++ Sources/MapboxNavigation/CarPlayNavigationViewController.swift
@@ -57,9 +57,9 @@ open class CarPlayNavigationViewController: UIViewController {
     }
     
     /**
-     Controls whether dark style will be used whenever traversing through a tunnel. Defaults to `true`.
+     Controls whether night style will be used whenever traversing through a tunnel. Defaults to `true`.
      */
-    public var usesDarkStyleWhileInTunnel: Bool = true
+    public var usesNightStyleWhileInTunnel: Bool = true
     
     /**
      Controls the styling of CarPlayNavigationViewController and its components.
@@ -524,7 +524,7 @@ open class CarPlayNavigationViewController: UIViewController {
     }
     
     private func checkTunnelState(at location: CLLocation, along progress: RouteProgress) {
-        guard usesDarkStyleWhileInTunnel else { return }
+        guard usesNightStyleWhileInTunnel else { return }
         let inTunnel = navigationService.isInTunnel(at: location, along: progress)
         
         // Entering tunnel